pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Yulia <jilly-2005@yandex.ru>"]
readme = "README.md"
package-mode = true

packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
django-bootstrap5 = "^24.2"
psycopg2-binary = "^2.9.9"
flake8 = "^7.1.0"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
coverage = "^7.6.0"
dj-database-url = "^2.2.0"
pytest-env = "^0.6.2"
rollbar = "^0.16.3"
django-login-required-middleware = "^0.9.0"
django-filter = "^24.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest_env]
DJANGO_SETTINGS_MODULE = "task_manager.settings"